+++ GUI/hideGUI.py
@@ -1,0 +1,5 @@
+import common
+
+def hideGUI():
+    if not common.isEditing:
+        common.showGUI.clear()
